полностью выполнены все задания module7-task3
